Update patch set 1
Patch Set 1:

Yep, that is the current quick verify job.

We need to do some work and identify what we should run in order to verify commits for different installers. If we enable full fuel build now, it will hit all the commits coming into genesis repo and increase the verification times for commits for all the installers.

So, I propose to go ahead with this as is (building docker) until we make verification installer specific. (if fuel stuff comes in, do full fuel build, if non-fuel stuff comes in, skip fuel build.)

Patch-set: 1