Update patch set 1
Patch Set 1:

Just an addition; another discussion we had with Aric to have slave/lab specific credentials to give write permissions for Google Storage. Setting up permissions for gsutil is quite simple. 

The plugin might require some other way to do this - or it might be tricky to support different slaves/labs with the plugin.

But, we can evaluate it for different use cases we have.

Patch-set: 1