Update patch set 4
Patch Set 4:

I must say that I'm not too fond of positional/unnamed arguments since they are hard to understand/not obvious and open to issues.

But anyway, if that's the way you prefer, I have nothing more to say as long as we fix current issues.

Patch-set: 4
Label: Verified=0