Update patch set 1
Patch Set 1: Code-Review-2

(6 comments)

This story was not about the implementation but was about identification.

Some of the questions that need answers are:
- Do we want to enable full runs or just keep them disabled?
- How frequently we want to run daily jobs?
- How to adjust the scheduling?
- Where to upload and download artifacts correctly?

All these and any other questions that require answers need to be discussed with bgs people to see what they need.

Perhaps we can come up with a suggestion and ask feedback from bgs.

Patch-set: 1
Label: Code-Review=-2