Update patch set 1
Patch Set 1:

(1 comment)

I'm still not so keen on enabling virtual deployments in verify jobs since things will be queued up if a change comes in while daily jobs are running, potentially blocking others if change touches to genesis/common.

So if this is not so crucial, I propose to leave this until we have the PODs reconfigured.

You don't have to abandon this by the way, just leave the stuff there but disable the job and remove the virtual-deploy from verify job builder.

Patch-set: 1