Update patch set 1
Patch Set 1:

Not sure how the workspace part works.  I added a git scm to the JJB, but from looking at the workspace list in Jenkins I'm not sure how this gets pulled to the workspace with folders "foreman" as a subdir of genesis.  Aric do you know?

Patch-set: 1
Label: Verified=0