Moves lab-reconfig to execute on master
The problem is the lab reconfig includes a reboot of nodes in a POD and
since Jenkins slave is on one of those nodes, it will sever the jenkins
connection and fail the job.  Therefore we need to run the lab-reconfig
from master node, then execute the deploy on the slave when it is back
up.

JIRA: OCTO-109

Change-Id: I10724b95685f6dba2cd463e01ff7618de292c851
Signed-off-by: Tim Rozet <trozet@redhat.com>