Update patch set 1
Patch Set 1:

Hi Fatih,
That's actually what I coded up first.  However, I thought that it wouldn't work in the case where you wanted to just run deploy job.  The reason I figured it wouldn't work is because deploy job is set to execute on opnfv-jump-2 as the host then would trigger lab-reconfig on master which I thought might have issues if anything was done on the opnfv-jump-2 prior to the reboot (git cloning, etc).  This may not be a factor at all and I can switch it back to the way you mentioned if you are confident it will won't have any repercussions.

Patch-set: 1