Update patch set 1
Patch Set 1: Code-Review-2

(2 comments)

Please explain the reason for this change and reconsider whether it is needed or not. 

If I'm not mistaken, not specifying file paths will cause triggering of compass-verify and compass-merge jobs for all the changes, including unrelated ones, that are coming to genesis repo, potentially blocking the work.

Patch-set: 1