Inclusion of 2 qtip_dhrystone jobs for the Dell Lab
First attempt at modifying/creating a JJB for a test case.
Used guide available at OPNFV wiki and JJBs for other teststing projects such as yardstick as a reference.
To be run on the dell-build node. If works, a job would be made for the LF POD too. Modified to remve white spaces and removed prebuilds

JIRA:QTIP-7

Change-Id: Iffed74a8bae64c4643839c4d2e2cac49a0070d9d
Signed-off-by: Nauman Ahad <Nauman_Ahad@dell.com>