Update patch set 3
Patch Set 3:

Hi I reverted your patch. The reason is that it passed verify but failed on the merge job. I have updated jenkins-job-builder, it now has proper checks for your code. please resubmit this patch in a new change set and it should fail verify. we can work on it from there.

Patch-set: 3