Update patch set 1
Patch Set 1:

Hi Fatih
We are trying to streamline the building of the vsperf documentation to what is recommended by opnfv. so we migrated all our documentation from markdown to restructured text and followed the instructions https://wiki.opnfv.org/documentation/tools, which is why/where this commit comes from. I don't think this needs to be part of the pre-commit gates that build our patches, but we do need a recommendation on how/where to create an appropriate job for building our documentation

Patch-set: 1