Prevent triggering unnecessary genesis jobs when docs change
Genesis verify and merge jobs run full builds when only doc updates come in.
This change adds non-doc paths to the list of excluded paths in gerrit trigger
configurations, essentially only running what needs to be run.

Please note that if a change touches both allowed and excluded paths, none of the
jobs will be triggered. Due to this doc and code changes should not be done in same
change.

JIRA: RELENG-19

Change-Id: I0388b8fb940aa7581e9afb56ffa04ca3a3763820
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>