Update patch set 20
Patch Set 20: Code-Review+2

This seems to be working as expected. I can start and stop connection using monit <start|stop> jenkins and the connection is established as jenkins user. 

It starts the process when I kill it manually as well so this is good to go in my opinion.

The only thing I want to get some info is about the change of monitoring status. I've seen 3 status messages; Does not exist, initializing, and running. I suppose the status changes between these states is not significant as long as monit does what it's supposed to do and keeps the process running.

Here is the log.

http://hastebin.com/eparimacik.mel

Patch-set: 20
Label: Code-Review=+2