Update patch set 6
Patch Set 6:

If you just want vping, I think there must be a way to not duplicate code, because it might lead to problems. For example, I have changed the cleanup builder to correct a bug, and you still have the old version... 
Can you try to call functest-vping without having to write the builder in this jjob?

Patch-set: 6
Label: Verified=0