Pass public key to docker to be able to access the Instack VM without password
JIRA: FUNCTEST-69

We need to fetch de openstack credentials from the Installer VM (Instack).
Apex leaves a SSH public key in the jumphost. The idea is to pass it to
the docker container so that we can access that VM from it without password.

This is only needed by Apex installation. Ideally, all should have the same,
so that we don't hardcode passwords.

Change-Id: I5e9f6c8f298140f77780362e640d43e624897a9c
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>