Update patch set 6
Patch Set 6:

I have only 1 comment left; I think we need to name jobs differently from the rest of the jobs in CI to make sure we now which one is used for development activities or local work, such as this one.

Will review it again once it gets +1 from jenkins.

Patch-set: 6