Update patch set 7
Patch Set 7:

I've been chatting with Narinder and he says they're waiting for this work to finish.

I must say whatever work is done in scope of this change will not work for CI. You use everything that is locally stored in your jumphost, you have references to home directory, you have parameters pointing to home directory.

Sorry to say this is not useful. We essentially should be able to use all these scripts for CI as well but they won't work in this state.

Patch-set: 7