Update patch set 7
Patch Set 7:

(7 comments)

As I said to Fatih by chat, I know that this actual config is not working, it is still WIP, and that is the way to share it with you. Thanks for your comments.

Those files are inspired by the fuel job (maybe too complex to understand by a jenkins newbie) and by https://wiki.opnfv.org/octopus/jenkins_wow. If you have another documentation page, please provide it to me.

The last point is that I need a way to have deployment passwords not shared on the internet. For me the best way is to let the admin pod set them in a local file in the jumphost. That is why I base my script on folder in $HOME (manual joid installation is also based on the home folder).

Other comments are in files

Again, thanks for your comments and the time you passed to understand my proposal.

Patch-set: 7
Reviewer: Gerrit User 361 <361@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0