Update patch set 7
Patch Set 7: Code-Review-1

Thanks for clarifications Daniel.

I tried to understand where the things stand with joid and wanted to make it visible that the work you are doing solves certain things but will not be enough for joid to become part of B-release. That's why I tried to make this obvious.

And please continue doing the good work.

Patch-set: 7
Label: Code-Review=-1