Fix compass jobs to stop using hardcoded branch name
Some jobs use hardcoded branch name master which will cause issues
when we branch for brahmaputra. In order to use var for branch name
some jobs appended stream.

Please note that this will result in removal of the jobs without stream
in them and creation of new ones, causing you to lose logs.

Change-Id: I65b498b189941f7137786c5d230373c7f5ecd492
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>