Update patch set 5
Patch Set 5:

Thanks for the reviews.

I chatted with Tim yesterday and he didn't oppose the idea. So this is good to go. I'll rebase and submit.

We can try this couple of days for joid and fuel and then apply to compass and apex. I'll be sending patches for compass and apex for you to review so you don't lose focus from your work.

Patch-set: 5