Enable generation of composite docs for opnfvdocs
Please note that the upload of documents to gerrit has not been enabled yet
in order to get feedback for the patch first. Once the patch is reviewed,
uploading generated docs to artifacts.opnfv.org can be enabled.

Another thing is that this could essentially be done in existing opnfv-docs
jobs. But in order not to break doc stuff totally, it's kept separate for the
timebeing.

Depending on the review, we can either
- move this functionality to releng-macros.yml and jobs in opnfv-docs.yml.
or
- keep it separate as it is and exclude opnfvdocs from the pattern in
opnfv-docs.yml jobs.

JIRA: RELENG-38

Change-Id: I452547921ef56960330bcd9ba28159108a33c860
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>