compass: Align scenario short names between installers
The fields in scenario short name is
    os-[controller]-[feature]-[mode](-[optionalparameter])

The change in scenario naming requires adjustment in jenkins job names
as well in order to have full alignment on CI.

The details of scenario naming scheme has been discussed during daily
standup on 2016-01-13.

You can read more details regarding scenario and job naming on
https://wiki.opnfv.org/brahmaputra_testing_page#scenario_and_jenkins_job_naming_scheme

fuel and joid switched to this type of scenario and job naming already.

Change-Id: Ia9246ac7ce8db054c1cb73381f83387ecd1dce66
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>