Update patch set 1
Patch Set 1:

Thanks for comments Thanh. Will remove the author names and just leave company names and the rest.

About SPDX, I don't think we have it for the rest of the OPNFV projects and I have no idea if it should be added.

Added Ray so we get some clarification for it.

Patch-set: 1
Label: Verified=0