Update patch set 1
Patch Set 1:

I'm aware of that this is still work in progress...

but I think we should have a quick chat before you put too much effort into this. there are some things that are slightly outdated in job setup you're basing your work on.
apart from this, you probably need some hw/bm to run stuff on so we can talk about this as well.

Patch-set: 1
Label: Verified=0