Update patch set 2
Patch Set 2: Code-Review+2

Ok. But to be sure, what I would do to avoid depending on the name, is to run a neutron command inside your container to get the public net name and export it as an env variable. 
We do something similar in functest (we ask neutron evey time we need to use it)

Patch-set: 2
Reviewer: Gerrit User 459 <459@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+2