Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

I think it makes sense to enable all these scenario once the release freeze is over. 
Today the priority is to get the 4 successful runs on the 3 joid existing scenarios, we probably should configure Orange POD2 to run these 3 scenarios in a CI way to reach these 3x4 successful runs.
We should also probably look at IMS test case which is still failed (not a problem from a release point of view but it is the last installer that does not support it. 
Ideally the 4 run should be done in Intel POD5&6 but I still do not have access (VPN allows me the connection to Intel POD 7 and 8)
Regarding the scenario, the wiki page shall be updated and the options collectively accepted by the installers in genesis.

So my suggestion is to wait B release to merge this patch

Patch-set: 1
Reviewer: Gerrit User 335 <335@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1