Update patch set 1
Patch Set 1:

> I started typing comments and ended up a long list so I moved them
 > to etherpad where we can have discussions as well.
 > 
 > https://etherpad.opnfv.org/p/cperf-ci
 > 
 > Please take a look at it and let me know of your comments/questions.
 > I hope what I've written on the etherpad makes sense.


Fatih,

thanks for all the explanations and help.  I actually continued the etherpad with
some questions and comments.  No problem to build on it there, if it still works
for you.

I think in the end, I'll push something wildly different on this gerrit for the
original purpose of getting an Initial CPERF job working.  The biggest problem
right now is time.  I'm only giving a minuscule effort to CPERF at the moment
and the Beryllium release is top of mind.

the CPERF pod is coming online (intel 2) and hopefully we'll start to get 
things flowing before long.

Thanks again for the support.

JamO

Patch-set: 1