Update patch set 3
Patch Set 3:

We can't keep changing the released artifacts. The link on software download page should point to one and only ISO that was released last week until we release the new version.

but it is good idea to have the latest ISO copied as opnfv-latest.iso for other reasons.

Patch-set: 3