Update patch set 3
Patch Set 3:

I'm not fully sure I understand this change. Only one build can be announced as the formal B-release ISO, so to me that would assume a manual operation? As that ISO is created it may never change - that ISO would be static until the next service release.

I would also expect the name of that ISO to be unique and easily identifiable as "the B release ISO", not "opnfv-lates" - and the SHA1 of that ISO would of course also be published on its download page, but that's of course obvious.

I'd appreciate if you could give some more info about the workings of this patch!

Patch-set: 3
Reviewer: Gerrit User 203 <203@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0