Update patch set 3
Patch Set 3: Code-Review-1

Aric, just saw your entry in the Fuel channel. No, we can not and should not go with latest. The build to release should be that one on which the release tests has been run - and the identity of which I think you've got in an e-mail from Jonas? We should not release anything else than what we've tested I believe.

I'd have to -1 this change as I understand the intention now.

Patch-set: 3
Label: Code-Review=-1