Update patch set 3
Patch Set 3: Code-Review-1

I agree; this is useful for other purposes.

BUT

There is something else that makes me think; for fuel, the ISOs produced by merge jobs are also copied to GS and the added code doesn't check which ISO to copy as latest. It should be the ones that are produced by periodic/daily jobs, not the merge jobs.

I suggest we skip this change now as we have plans to abandon periodic builds and use ISOs built by merge jobs by storing them on GS properly.

Patch-set: 3
Label: Code-Review=-1