jenkins-jnlp-connect.sh: monit config file was always updated
The jenkins-jnlp-connect.sh script will check whether a monit
configuration file for jenkins exists. If it does, it will make sure the
contents are correct according to the paramters provided in the command
line. The check is made diffing the contents of the file with what is
expected by the script. However the script compares the contents of the
file against a string that is not exactly the same to what would be
written to the file by the script, resulting in the check always
failing, and the file being rewritten unnecessarily every time.

Change-Id: I0b15195e9d667c11898942c92b39ef3fb053821d
Signed-off-by: Josep Puigdemont <josep.puigdemont@enea.com>