Update patch set 4
Patch Set 4:

(1 comment)

I suppose the work with this patch is still going on so please let me know when you want this to be reviewed since I have some comments/questions and answers perhaps.

But I must say this is much better than previous patches; as I mentioned, separating the deployment stuff from test projects is the approach we follow for other test projects as well.

Patch-set: 4
Label: Verified=0