Update patch set 2
Patch Set 2: Code-Review-2

need to block this.

we need to change slave name on jenkins to nokia-pod1 since this is not a build slave and reconnect it back with a new token. I'm around this afternoon so ping me on irc when someone is on the jumphost. (irc handle is fdegir)

Patch-set: 2
Label: Code-Review=-2