Update patch set 7
Patch Set 7:

(4 comments)

another try to reduce the no of questions even further.

But please do not think I'm getting bored or something and keep coming with the any questions you might have so I can learn more if what we do makes sense for you and improve as needed.

In the end, you will receive a questionnaire with 125 questions to evaluate your experience. :)

Patch-set: 7