Update patch set 2
Patch Set 2:

OPNFV is a bit different than ODL so my opinion on this is likely different than yours. In our experience we actually prefer to not allow people permissions to Jenkins at all. On the ODL side of things we're looking into providing less access to Jenkins itself as we move forward with future CI changes.

The problem with Jenkins as we know is that it has scaling issues. Including too many people logging into Jenkins and navigating the UI can cause it to become unresponsive because the UI itself is single threaded. The kind of workflow being proposed here seems like it will encourage more people to access the Jenkins UI and you will likely run into scaling issues in the future due to this if you're not already experiencing slow Jenkins responses yet.

Patch-set: 2
Reviewer: Gerrit User 245 <245@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0