Update patch set 2
Patch Set 2:

Thanks for the feedback Thanh. I agree with you about limiting the access but the nature of OPNFV requires us to provide some type of "as a Service" approach. That's why we have Jenkins access.

Apart from this, we already provide access to people and this proposal tries to limit that access.

There are number of different reasons why we had to provide Jenkins access to people.
- PODs/slaves coming from many different labs. It is pretty time consuming for people to get access to this many labs so Jenkins is one way for people to access to the labs.
- We do not want people to mess with the PODs by directly logging in to them as any change in POD config might cause issues.
- We have many different installers, test projects, feature projects and so on. It is not easy for people to know the details of all these different pieces, put them together and do the installation and testing. Again, Jenkins gives them an easy way to deploy their stuff in a controlled manner.
- We want to make sure people do stuff out in the open using the utilities provided by different OPNFV projects as automatically as possible. (starting Jenkins Job manually is nothing automatic but at least we know they use the scripts provided by other projects so OPNFV installation and testing is done automatically.)
- Above item helps others who might need to troubleshoot stuff for the users. With Jenkins we again know how people do what they do and so troubleshooting becomes easier.

Long story short, the Jenkins access should be seen as a service to the community to ease the onboarding process and help with the learning which is pretty steep.

Patch-set: 2