bottlenecks: Disable jobs running against CI PODs
Test Strategy Group must come to a decision regarding what to run against
CI PODs. Until that happens, business as usual; deploy, functest, yardstick.

In order for bottlenecks to run against CI PODs, it must be discussed and a
decision must be presented.

If the jobs get activated and run again, next step is removal of the jobs.

Apart from CI PODs, permission was never asked and never given for ericsson-pod1 so
job running on it is disabled as well.

Change-Id: I82b481775475ef36d132541efdc40c325d05e77a
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>