Update patch set 2
Patch Set 2:

Just a clarification; the issue is not specific to bottlenecks. It is more general question.

For example, qtip came with same request and they've been asked to do the same thing; until Test WG comes to a decision, qtip needs to be run against none-CI PODs. qtip moved to none-CI PODs and waiting for decision to be taken.

But the situation with bottlenecks is different and has become more serious as all the attempts failed to convince the project.

Patch-set: 2