Update patch set 2
Patch Set 2:

While we are discussing this patch, you run the job again about an hour ago.

https://build.opnfv.org/ci/job/bottlenecks-fuel-rubbos-opnfv-jump-2-daily-master/

And this is the history for ericsson-pod2 that is used for plugfest. The job was run on Monday, just before the plugfest. It's not been run since Monday because I had to disconnect the POD from Jenkins in order to protect it. We shouldn't really need these things. It is pretty disappointing.

https://build.opnfv.org/ci/job/bottlenecks-fuel-rubbos-ericsson-pod2-daily-master/

Patch-set: 2
Label: Verified=0