Update patch set 3
Patch Set 3:

The POD opnfv-jump-2 is not reserved for plugfest! It is used for CI and only deploy, functest, and yardstick are allowed to run on it!

Would you like to remove -2 or do you want me to abandon my patch and send a new one. 

This is not really going to the direction that would benefit entire community.

Patch-set: 3