Update patch set 2
Patch Set 2:

(1 comment)

I can imagine that we need this type of control mechanism as we have multiple organization connected to the jenkins. And, I understand setting permission per POD. But, why we set it in job definition? Is that the only or preferred way to tell the permissions to slaves?

Patch-set: 2