Update patch set 1
Patch Set 1:

Also, is this .yml missing "triggers"?  I think that is one thing
missing, else how does the job run.

another question (I can ask on the mailing list if needed) is how we
can manually start our jobs?  In ODL we have permission (just committers to the projects) to manually (build) our jobs.  This is not
an option for me in OPNFV jenkins.

Patch-set: 1