download/deploy scripts: several updates
The download script will now use gsutils if they are available, or fall
back to curl as before.

CHANGE: A timestamp will be appended to the ISO image file name, this
name is then stored in latest.properties, so the deploy script will be
aware of it after it has sourced the file.

In general, use curl -f, which will exit with an error code if the
server gave an error, instead of writing a file with the HTML error
page.

Change-Id: Ib24e7803ec38b20a8879cd1509a0cb7b512fee61
Signed-off-by: Josep Puigdemont <josep.puigdemont@enea.com>