Update patch set 16
Patch Set 16:

Some comments.

Having branch specific files is not really sustainable if there is other way to fix it. 

Apart from this, we should try to keep on having as less duplication as possible and keep the no of templates as low as possible to increase the reuse and reduce the maintenance effort.

This also increases the risk of divergence.

Please take a look at and put your comments on my patch as well.

https://gerrit.opnfv.org/gerrit/#/c/14159/

Patch-set: 16