Update patch set 6
Patch Set 6: Code-Review-1

This patch is just the first try to see how we can handle CI and none-CI PODs together.

It is not a good idea to do all these changes in one go so this will be split to multiple patches - per installer perhaps and the rest needs to be updated in subsequent changes.

Another thing to keep in mind is that we can perhaps keep the slave based jobs and create additional baremetal jobs to see how it goes. If it works fine, we remove the slave based jobs and continue with label based ones.

Patch-set: 6
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1