Update patch set 16
Patch Set 16:

I see apex, compass, fuel, joid, all same and I think we should try everything possible to keep them aligned, as much as possible. This might introduce limitations but helps greatly to keep things aligned.

It is not just about alignment either; think about others who are working with all installers. (like test projects) If they want to do something for all installers, they will need to put different stuff for different installers. (this problem we already have due to apex.)

So, I say no to having this for compass and having the other for other installer. It is not about the solution or anything else. If we do something, we need to do that for all not just one.

Finally, some limitations you highlighted do not exist; they can be done in different ways in the other patch.

For example testing scenario on one branch; One solution to this is disabling trigger for that scenario on that branch using brahmaputra-trigger-daily-disable. Or enabling with baremetal-brahmaputra-scenario-trigger etc.

Again, you can disable triggers for none-ci pods for different branches, this is possible as well.

Patch-set: 16
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0