[docs] improve config
A developer was asked to copy all contents in the base config file in
order to set all parameters properly, even if the change is partial.

This patch improves this config by setting all default parameters based
on users' custom config file, so that they can write parameters just
which they want to change.

Change-Id: I963fc5e3fde79abe87e2ed9f54f496f9bdce31df
Signed-off-by: Ryota MIBU <r-mibu@cq.jp.nec.com>