Update patch set 1
Patch Set 1:

hi, fatich and jose, thanks for your great suggestion. As Jun said, there are two dockerfile and two run_test.sh in yardstick. Sometimes, it makes users confused. So Jun proposed that we should adjust the directory of yardstick to make it clear. As the project inspired by Rally, we make rally's directory as reference. However, we missed to make the alignment with the rest of projects as a consideration. thank you for your kindly reminder. but i saw there are two location in the opnfv-docker.sh for different projects. do we have a plan to make them aligned

Patch-set: 1