apex: Take builders out of jjb
Having builders embedded in jjb makes maintenance harder than what
it could be and error prone so the builders have been taken out.

Apart from this, the upcoming change proposals will split yml files
in 2, aligning with the rest and jobs in these files can reuse same
builders when they are converted in shell scripts.

This is the first change in series, trying to align job structure
with the rest in order to ease the effort required to troubleshoot
the deployments during release verification. The changes are
- take builders out of jjb (this patch)
- align the main CI jobs with the rest; 1 parent job per scenario/branch
  which controls triggering and execution of 3 downstream jobs;
  deploy, yardstick, functest
- split yml file in 2; ci jobs ending in apex-ci-jobs.yml (daily) and
  project jobs ending in apex-project-jobs.yml. (verify, merge, build)

Change-Id: I01f8bf351f0cbafbee1f41ec24ad5c2c6f21316f
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>