Update patch set 2
Patch Set 2:

we did not consider this as a parameter in previous versions.
As far as I understand, it will now be possible to select the OS as one of the settings when building your OPNFV scenario.
shall we then imagine to run tests on the same scenario with different OS?
is the influence of the OS big enough to justify a dedicate scenario?

As far as I remember, the requirement was to support both Ubuntu and CentOS
this changes the agreed rule on the scenario naming
os-<controller>-<feature>-<mode>
could-it be considered as a feature?
shall we modify the rule?

Patch-set: 2
Reviewer: Gerrit User 335 <335@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0