Update patch set 2
Patch Set 2:

That's the question we need an answer for.

If it adds significant value, it can be part of the scenario, just highlighting the fact that the deployment is done using that certain OS. 

Apart from this, the scenario naming scheme supports this since we have the optional field in the scenario naming. The actual naming is like this

os-[controller]-[feature]-[mode]-[option]

No matter what we decide, we need to keep things aligned; if we go this way and include OS in above scenario naming, we should do this for all meaning that option becomes mandatory no matter if they support both OS or only one.

And it should be like ubuntu14, centos7 rather than using trusty for ubuntu and centos7 for centos.

https://wiki.opnfv.org/display/SWREL/Brahmaputra+Testing+Page

Patch-set: 2