Update patch set 3
Patch Set 3:

Regarding slave names in jobs; yes and no.

For the jobs that are running on CI PODs: slave names are not in job names anymore. They are named as [installer]-[scenario]-[baremetal/virtual]-daily-[branch]

But for none-CI PODs, we need to continue using slave names in job names as we (might) need to handle them differently and make it visible that the jobs running on those PODs are not taken into consideration for release purposes.

Patch-set: 3