Update patch set 4
Patch Set 4: Code-Review-1

If we go down the path of adding versions numbers to jobs we should do so consistently - and with the appropriate level of detail. I.e. Centos7 could be Centos7.1 Centos7.2 Centos 7.25.., etc etc. Similar things are true for OVS, KVM, etc.
IMHO the scenario name is not to replace a config file - and we're seemingly on a path to turn the name into a config file here.

Patch-set: 4
Reviewer: Gerrit User 251 <251@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1