Update patch set 4
Patch Set 4:

I agree with what Frank says since the idea with having scenario names in job names to make the scheduling and browsing on jenkins simpler.

We need to find a different and sustainable solution for this and keep the scenario names in jobs as simple as possible, solving our real jenkins issue.

May, 

can you come up with a different proposal please? Perhaps a POD specific parameter in slave-params.yaml or whatever that tells the OS and whatnot and that info get logged on Jenkins console when the job runs for visibility.

Patch-set: 4
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0