Update patch set 3
Patch Set 3: Code-Review-1

I don't have a problem with cperf jobs consuming/triggering apex jobs on their own separate pod.  That is good re-use of the jobs we already define in apex.yaml.  The question that determines if that is the best path is: How do you plan on running cperf?  If you are OK with triggering once daily, then we can run it for you as part of apex-daily.  If you want your own control and be able to run it as a cperf daily, or whenever you do your verify, the job should be in cperf and trigger an apex deployment.

Patch-set: 3
Label: Code-Review=-1