Update patch set 3
Patch Set 3: -Code-Review

Tim,

First of all, apologies for a long response but I must explain things in more details here so everyone knows the reasons. 

I think we always end up in same type of discussions which I don't complain having them. It means we have healthy relationship; we can discuss things. 

The subject of our discussions is generally related to alignment stuff; slave name alignment, way of working alignment, job structure alignment, pod usage alignment. I agree you have your own projects, you have your own pods and so on but you are part of something bigger at the same time. For your project, you can do anything but if you work with others, you need to follow common ways.

When it comes to this patch and Cperf in general; Cperf will still keep their own stuff in their jjb and whatever code they have will still be in cperf. 

As test project, Cperf shouldn't really touch the deployment parts so I don't see a reason for them to change apex.yml or the apex code; they just need a standard OPNFV scenario deployment which will initially be provided by Apex and later on by other installers if Cperf has plans to do that. If Cperf introduces changes to Apex stuff then they will be doing something else then what Apex does for the rest of the community. This might be the case and on purpose, I don't know. But I can only provide my comments based on what I know about Cperf; it is one of the OPNFV test projects focusing on controller performance.

About the only way or not; I agree there are tens of different ways of achieving same thing or even better things. But this does not mean we all go and do our stuff on whatever way we like. We need to have alignment in order for other people to make sense of what we are doing. We need our users to do things in common way. This was an issue for everyone during Brahmaputra release phase and still is an issue; things are different and people have to try hard to understand what Apex does and what others do. This is especially bad for test projects.

An example for this; dockerizing stuff; Cperf could have gone to its own way and do things whatever way they like due to different reasons. But this could have impacted our end users who want to use Cperf. They pull docker images for all the test projects but do something else with Cperf. Same with CI, Cperf can be run on any POD Jumphost which has Docker installed rather than installing all the tooling and dependencies.

Long story short and I know I repeat myself but I must say this; we need alignment and what I explained here is obviously not the best way. Someone could come up with better way and we all go that way. What I explained here is the current way used by the majority of the community. I will continue pursuing alignment all the way as much as possible to make lives of our developers and users a bit easier - which might make your lives harder in the end but that's the price we need to pay time to time.

So I'll not block this in order for Cperf to continue with what they want to do based on the way you all agree. But Cperf will need to go that way one day either for other installers or when you change the way you do things for Apex.

Patch-set: 3