Update patch set 3
Patch Set 3:

Thanks for the explanation.  I guess the part I don't get is if there is some standard way of doing things -  was it formed because functest/yardstick did it this way, or did some governing body decide the standard?  Is it up to releng/octopus to decide on such a standard?  If so, did the comitters hold a vote?  If not, we should vote on a standard way, and form a document for new tests projects on how to structure their JJB.

Patch-set: 3