Update patch set 3
Patch Set 3:

I think in general it is good to standardize jjb jobs but this standard should as a minimum fulfill most projects requirements and scope. In the context of the cperf project, we will try to summarize and communicate all things we need and see if current or future standard supports that.

Patch-set: 3
Reviewer: Gerrit User 3717 <3717@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0