Update patch set 6
Patch Set 6:

patch 5 addresses Tim's comments, but patch 6 is my weak attempt to 
steer it back toward Fatih's vision.

I understand fully about trying to keep some sort of standard way of 
doing things to make it easier for others (especially new people) as
we deal with it constantly in our ODL Integration/CI groups.

I am so short on bandwidth that I rarely have time to spend on this,
so please don't take my absences as apathy.  I am trying to understand
how to do it, but I am almost shooting in the dark here.

I want apex to deploy opnfv on intel pod 2 and then I want to run the
stupid "hello world" cbench tests against that.  there are two builders
in the cperf-ci-jobs.yml for that.  They are going to fail and I have to
figure out all the other details to get it right.  That's the place I
want to be.

That's all I'm going for here.  This should be typical enough for test
projects  right?

Patch-set: 6
Label: Verified=0