Update patch set 6
Patch Set 6:

On further review, it appears that Fatih is suggesting that we do not modify apex.yml but instead import jobs from apex.yml in our cperf-ci.yml. 

"When it comes to this patch and Cperf in general; Cperf will still keep their own stuff in their jjb and whatever code they have will still be in cperf. As test project, Cperf shouldn't really touch the deployment parts so I don't see a reason for them to change apex.yml or the apex code; they just need a standard OPNFV scenario deployment which will initially be provided by Apex and later on by other installers if Cperf has plans to do that. If Cperf introduces changes to Apex stuff then they will be doing something else then what Apex does for the rest of the community."

Can you confirm my summary is correct Fatih? If so, it implies abandoning this patch and continuing work in cperf-ci.yml.

Patch-set: 6
Label: Code-Review=0