Update patch set 6
Patch Set 6:

Marcus & Jamo

The stuff we do on Jenkins is mainly driven/controlled by installer jobs. This means that upstream jobs are created in installer jjbs and these upstream jobs call different jobs from other test projects. The test project jobs get defined by the test projects and then run by installer jobs. The reason for this is that all the scenario definitions, triggering, pods, etc. are only/mainly in installer jjbs and have the clear separation on how is doing what.

By putting cperf job triggering in apex.yml, you basically are doing what we do for the rest of the test projects.

Let me know if this answers some of the questions you might have. There are tons of other details as you have seen in previous comments and there are some other changes in the pipe but we can skip them for the timebeing.

Patch-set: 6